requirements.txt
aiogram==2.25.1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.12.1

